requirements.txt
matplotlib==2.0.2
numpy==1.13.3
Pillow==4.3.0
scikit-image==0.14.1
simplejson==3.8.2
sklearn==0.0
tabulate==0.8.2
tensorflow-gpu==1.9.0
